backend/pipeline/requirements.txt
polars>=0.20.0
pandas>=1.5.3
dagster>=1.2.7
dagit>=1.2.7
dagster-postgres>=0.18.7
polars>=0.20.0
pymongo>=4.5.0
minio>=7.1.15
langchain-google-genai
langchain-core
pyarrow>=11.0.0
pydantic>=1.10.12
python-dateutil>=2.8.2